+++ util.h
@@ -1,11 +1,11 @@
 #ifndef UTIL_H
 #define UTIL_H
 
-static const int pow2[] = {1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048, 4096};
-static const int pow3[] = {1, 3, 9, 27, 81, 243, 729, 2187};
-static const int fact[] = {1, 1, 2, 6, 24, 120, 720, 5040, 40320, 362880, 3628800, 39916800, 479001600};
+static const long long pow2[] = {1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048, 4096};
+static const long long pow3[] = {1, 3, 9, 27, 81, 243, 729, 2187};
+static const long long fact[] = {1, 1, 2, 6, 24, 120, 720, 5040, 40320, 362880, 3628800, 39916800, 479001600};
 
-static const int pick[][NUM_EDGES+1] =
+static const long long pick[][NUM_EDGES+1] =
 {
     {1,  0,   0,    0,     0,     0,      0,       0,        0,         0,         0,         0,         0},
     {1,  1,   0,    0,     0,     0,      0,       0,        0,         0,         0,         0,         0},
@@ -22,7 +22,7 @@ static const int pick[][NUM_EDGES+1] =
     {1, 12, 132, 1320, 11880, 95040, 665280, 3991680, 19958400,  79833600, 239500800, 479001600, 479001600},
 };
 
-static const int choose[][NUM_EDGES+1] =
+static const long long choose[][NUM_EDGES+1] =
 {
     {1,  0,  0,   0,   0,   0,   0,   0,   0,   0,  0,  0, 0},
     {1,  1,  0,   0,   0,   0,   0,   0,   0,   0,  0,  0, 0},
